bindings: Checks that non-preemptible sizes do not have "preemptible" flag set
Before this commit, the chooseSize function did not distinguish
between spot and reserved machines. It simply assumed that all
machines can be preempted. This is not always the case, e.g. Amazon
t2.micro can only work when using reserved machines. This commit
makes sure that the user is using a valid preemptible size if
preemptible is set to True, and errors otherwise.