digitalocean: Don't attempt to create firewall if it already exists
This fixes a bug where the DigitalOcean client would only request the first
page of the firewall listing, causing it to potentially not find the
existing firewall for a namespace and attempt to recreate it, which would
fail and cause all of the ACL related API calls after that to fail too,
preventing the daemon from ever actually connecting to the machines.

Practically speaking, this should fix the integration tests for DigitalOcean.