travis: Fix NODEJS_VERSION deployment check
The check didn't actually evaluate the NODEJS_VERSION variable in the
conditional. In fact, it assigned to the variable.

This commit also changes the check to use `==` rather than simply `=`.
Although both are valid with the `test` command in bash, `==` is more
consistent with the rest of our travis.yml, and easier to read.