cloudcfg: Create TLS credentials directory before minion runs
Before this commit, there was no guarantee that the minion's TLS
directory would exist before it started. If the directory does not exist
when the minion starts, Docker autocreates the directory as root.

This resulted in a race where if the daemon did not install the minion's
credentials before it started, Docker would autocreate the credentials
directory, and the daemon would thus be unable to write to it. Then, the
minion would stall indefinitely because it never received its TLS
credentials.

This commit avoids this scenario by creating the TLS directory with the
proper permissions in the cloud config.