travis: Ignore node_modules when creating test releases
Before, the tarball that Travis created for Jenkins to use in the
integration-tests contained the node_modules directory. This resulted in
the `npm install` done during the integration tests being a no-op because
the packages installed during the Travis build were already present, so
the `npm install` run by Jenkins does not think it has to do anything.
Therefore, any updates to the Node packages after the release was
created did not get picked up by Jenkins.