scheduler: Wait for Vault before attempting to boot containers
Before, the workers would sometimes log scary errors about being unable
to connect to Vault. This patch makes it so the worker first ensures it
can successfully connect to Vault before attempting to boot containers.
This way, if we fail to connect to Vault later on, we know it's a real
issue, and deserves to be logged at ERROR.