travis: Fail build if unable to push to DockerHub
There were several issues with the old method of pushing images to
DockerHub:
- Rather than using the `deploy` target, we used an `if` statement in
the `after_success` target. Errors within the if aren't propagated, so
even if the push failed, the `if` would still exit 0.
- The commands were chained with `;` rather than `&&`, so even if the
first command failed, the next one would still execute, and the exit
code would be that of the final command.
- We pushed regardless of whether we were building the
DEPLOY_NODEJS_VERSION, so two images would get pushed for every commit.