tests: Refactor the network test
Before this patch, the network test code was extremely complicated and
slow.  SSH attempts tended to come in batches, causing the system to
either bottleneck on the concurrency limit, or waste time trying to
figure out what to do next.

This patch refactors the network test and in the process dramatically
simplifies the code. By reusing SSH connections and removing
sync.WaitGroup barriers, the network test speeds up dramatically,
taking about 45 seconds in my local setup.  Furthermore the test
scales much better to very large clusters.

This patch also causes the network test to print out ping and dns
attempts as they happen and failures as they occur, when running in
verbose mode.  This makes development easier as network failures can
be acted on and dealt with without having to wait for a full test run.