foreman: Generate etcd and public key configuration based on database
Before, we generated the etcd configuration and public key mappings
based on the `minions` map. However, when a minion disconnects, it is
removed from the map, causing the configuration that we send to the
minions to change. This can sometimes result in transient errors -- for
example, if a minion is removed from the public key mappings, then the
leader will revoke Vault access to it, causing the minion to log some
errors about connecting to Vault.

This commit avoids this problem by generating the minion config based on
the database, whose contents are not affected by the connection status
to the minion.