dns: Properly respond to IPv6 queries
Before, we responded to DNS queries for IPv6 addresses with the Not
Implemented code. This caused DNS resolution to be extremely
slow when running on the Google provider (about 3 seconds to resolve
google.com).

The proper response when there are no IPv6 addresses is to provide no
answers, as per https://tools.ietf.org/html/rfc4074. Resolving
google.com on the Google provider is now more or less instantaneous.