blueprint: Expose the deployment object through global function
Before, we exposed the deployment object directly as a global variable that was
used in blueprint.go to get the blueprint from a file on disk.

However, an upcoming commit removes the Deployment class in favor of the
Infrastructure class. Since the Infrastructure constructor requires a set of
Machines, we can no longer initialize the global _keldaDeployment variable to
an empty deployment. Thus, to handle an empty deployment in the future, we now
obtain the deployment object through a global function that wraps
`_keldaDeployment.toKeldaRepresentation()` and takes care of any uninitialized
deployment.