readme_test: Use bindings from source rather than latest release
Before, the README test pulled the latest Kelda release when checking if
the README blueprint compiles. This means that the test and thus build can
break if we make breaking changes to the JavaScript API (until these changes
are merged and released). To avoid this, we now instead `npm link` the
README test to the source code bindings, so the test runs against the latest
API and source, rather than the latest release.