acl: Allow return ICMP traffic
Commit 233b1a1c (acl: Avoid duplicate ICMP ACLs) forgot to migrate the
rule for allowing return traffic for ICMP. This resulted in `ping`s only
working when containers were explicitly connected in both directed in
the blueprint.

Note that this commit does not fully restore the previous behavior of
allowing either the source or destination to initiate ICMP connections.
Now, only the source container can initiate ICMP connections, and the
destination can only respond.