api: Query cluster information from cache
Before, every call to `kelda show` would cause the daemon to query all
machines for the state of the containers. This was inefficient,
especially at scale. With this commit, querying the cluster is separated
from retrieving the information for `kelda show` -- the daemon
constantly queries the cluster, and `kelda show` simply displays the
most recent information (which should be no more than 10 seconds old).

Another benefit of this change is that temporary issues to connect to
the cluster are hidden from the user. `kelda show` will simply display
the last successful call, and any errors communicating with the cluster
is logged in the daemon.