etcd: Persist data to host disk
Before, etcd data was saved to the container's disk, but not persisted
in the host. Therefore, if the etcd container was restarted, all the
information in etcd was lost. This resulted in a bug where if a new etcd
configuration was sent to the minions (for example, right after the
daemon restarted, or the number of masters was changed in the
blueprint), all containers would be restarted. This is because sending a
new etcd config triggers the etcd container to restart, which in turns
causes the minions to read that there are no scheduled containers,
resulting in it destroying whatever containers it was previously running.