tester: Fix Spark test to use driver for job
A recent change to the Spark repository changed the blueprint to
setup the spark-driver machine (instead of the spark-master machine)
with the correct permissions and environment variables to launch jobs.
This commit updates the Spark integration test accordingly (before
this commit, the Spark integration test ran a Spark job in local
mode, so it would run in a single JVM on the Spark master, rather
than distributing work to the worker machines).