tester: Properly count number of Spark workers
Before, the duplicate cluster treated all containers that weren't the
Spark master as workers. However, a recent change to the Spark blueprint
made it so an additional Spark driver container is deployed per cluster,
so the integration tests were failing because it treated the driver
container as a worker, and did not see as many worker connections as it
expected.