bindings: Set CPU and RAM attributes of Machine
It can be useful for application blueprints to know the
CPU and RAM of the Machine that will be launched. This commit
adds cpu and ram properties to the Machine object, and sets these
properties to be the exact CPU and RAM of the cloud instance
that will be launched.