init: Only allow a single base infrastructure
It wasn't clear why users would create multiple base infrastructures, and the
ability to do so added complexity to both the docs and init command. With this
commit, users can only have a single base infrastructure at a time. For backwards
compatibility, this infrastructure is located in ~/.kelda/infra/default.js.