server: Fix a race when changing namespaces
Before, when a user deployed a blueprint with a different namespace the
foreman would sometimes send the new blueprint to any existing machines
in the old namespace. This was because there was a period of time where
the blueprint was updated in the database, but the cloud join had not
yet removed the machines from the database. During this time, the
foreman would send the new blueprint to the machines in the database.

This commit fixes this by removing any existing machines from the
database in the same transaction as adding a new blueprint that changes
the namespace.