ci: Fix logic to wait until machines are connected
Now that the blueprint machines aren't written into the database, there
is a period after the blueprint is run where the machines table is
empty. Therefore, waiting for all machines in the database to be
"connected" is insufficient since the empty list satisfies this
constraint.

This commit adds an additional check that the number of machines in the
database is equal to the number specified in the blueprint.