ssh: Demote most SSH key errors to DEBUG
Kelda tries many locations for SSH keys; for most of those locations,
we don't expect to find a key that Kelda can use. This commit message
demotes all log messages associated with SSH keys to debug unless
they're for the default Kelda ssh key path, and adds a unit test
for this functionality.