ci: Speed up test compilation
The integration tests were taking multiple minutes to compile in the
Jenkins container because none of the test dependencies (which includes
the Kelda source) were cached. The build was fast when building locally
if `make install` or `go install` had been run recently.

This commit adds the `-i` flag when building the integration tests so
that any test dependencies are installed and cached, and can thus can be
used by other test builds (for example, the dependencies of 10-network
can be reused by 15-bandwidth).