scheduler: Properly set container's hostname
Before, hostnames within containers defaulted to a hostname generated by
Docker, which broke some applications that tried to resolve their own
hostname. This patch sets the hostname to be that specified by the
Container constructor in the blueprint, which can be used to access the
container from within the Kelda private network.