cloud: Make provider initialization more robust
This commit makes two related changes to the way providers are
initialized:
(1) Continuously re-try to initialize providers when initialization
fails, so that the Daemon can connect to a provider if, for example, a
user adds credentials in the correct location after the daemon has
already started. This functionality used to exist
(https://github.com/kelda/kelda/commit/dd11b5b060b8af08d183a29f3f2ca16c412848d1)
but was removed when the cloud code was recently refactored.
(2) Log an error message (rather than a debug message) when a provider
that is used by the current blueprint cannot be initialized. This
fixes a problem where, unless the user had debug logging on, a user
could run a blueprint that references a provider that isn't properly
configured, and the daemon would silently fail.