node: Try both node and nodejs binaries
Even though `node` is the blessed name for the Node.js binary, some package
managers (e.g. apt-get) install the binary as `nodejs`. Before, we only tried
the `node` binary, but with this commit, Kelda also tries the `nodejs` binary.

The `nodejs` name is used by package managers that already had a package named
`node`. To not accidentally execute the wrong `node` binary, we try the nodejs
binary first.