ci: Set blueprint namespace directly in Infrastructure constructor
Before, the integration-tester set the namespace of the blueprints by
appending a magical string to blueprints before running them. This
relied on undocumented parts of the Kelda API, and prevented tests
that shelled out to `kelda run` from running in the correct namespace.

This commit makes it so `infrastructure.js` reads the namespace directly
from the TESTING_NAMESPACE environment variable. This way, the namespace
will be properly set in any files that `require` the infrastructure
definition.