ci: Wait until machines are actually stopped
Before, the stop function just slept for two minutes, but we didn't know
whether the stop command actually successfully destroyed the machines.
Now that the daemon tracks all the machines that are actually running we
can directly check whether the machines have stopped by ensuring the
machines are no longer in the daemon's database.

Note that this commit still hardcodes a sleep of two minutes so that the
daemon has ample time to do the first list of the currently running
machines.