cli: Make 'kelda stop' prompt the user for confirmation
Previously, the stop command would immediately tell the daemon to halt
all of the machines or containers without confirming with the user that
this is the intended action. Now, the stop command warns the user if
running 'kelda stop' will terminate any containers or machines.

Some logic was slightly simplified as well, which caused 'stop' to not
bother deploying an empty blueprint if it was already empty, which had
the added effect of requiring some of the previously existing tests to
be slightly changed.