ci: Check daemon log output for errors
This modifies the 100-logs integration test to check the daemon log
output for errors in the same way it checks the minions' log output.

Specific types of log lines may be ignored by specifying a regex --
this is intended for benign errors so that they don't cause the
integration tests to fail while they're being addressed.