ci: Make setting up the infrastructure a test suite
Before, booting the machines that the integration tests would run on was
handled differently than booting the integration test blueprints. This
had several drawbacks:
- Any post-run debugging would need to be duplicated (most notably,
running `debug-logs`)
- It was hard to tell if the infrastructure failed to setup. Before, a
user would have to look in the "Console Output" of the Jenkins build, but
now the result will just show up in the "Test Results" view along with
all the other tests.
- It made the integration test runner more dependent on the structure of
the tests (specifically, we had to hardcode a path to `infrastructureRunner.js`
in the test runner source).

Note that because setting up machines is no longer specially treated,
the tests will not exit early if a machine fails to boot. However, this
might actually be useful since we will get to see how the tests behave
when a machine is missing from the cluster.