db: Send an initialization tick for Trigger
Our codebase assumes that database triggers will send an initialization
tick so that the code can perform an initial sync. This is necessary to
prevent race conditions where an event that would normally cause a
trigger happens before the Trigger is registered with the database.

This commit makes the Trigger function send an initialization tick.
TriggerTick already had this behavior.