cloud: Don't assign floating IPs until machine role is known
Before this commit, when booting a cluster with a floating IP, we would
essentially pick a machine with a random role to assign the IP. This was
because of the delay between booting a machine, and finding out the
machine's role.

This commit makes it so we don't attempt to assign floating IPs before
we learn the machine's role.