cloud: Prioritize role matches over floating IP
Before, the cloud join would sometimes prioritize matching machines with
the same floating IP over machines with the same role. This wouldn't
happen if the roles were explicitly different (Master vs Worker), but
there is a period of time when the role is unknown to the cloud, and in
that case. This resulted in a bug where machines would be unnecessarily
destroyed when there was a master and a worker, but one had not
connected yet.

Consider the following join:
In the blueprint:
- A Worker with a floating IP.
- A Master.

In the cloud:
- A Master-ToBe with the desired floating IP, but that had not yet connected
back to the daemon, so its role is "None".
- A Worker without a floating IP.

The old join would pair the blueprint Worker with the Master-ToBe since
the floating IPs matched. Then, the remaining machines would be unable
to be paired since they are a Master and a Worker. Therefore, the cloud
would destroy the Worker, and boot another Master.

This commit fixes this by making machines that have matching roles a
better match than machines that only have matching floating IPs.
Machines that have the right role and floating IP are matched even
higher to avoid arbitrarily reassigning floating IPs between machines
with the same role.

Note that although the previous commit that delays assigning floating
IPs until the roles are known should make it so this issue doesn't arise
when the machines are first booting, it was still a problem when the
daemon is started with machines already running in the cloud.