ci: Add check for lost tasks to Spark test
Sometimes, Spark tasks fail, but Spark recovers enough that the
integration test job succeeds.  Tasks shouldn't fail while the
test is running (this is typically an indication of some problem
with Kelda), so this commit updates the integration test to explicitly
test for these failures and fail if they occur.