google: Don't error if a machine doesn't have an IP
Before, if a machine entered a bad state where it didn't have an IP, the
Google provider would return an error when listing. This resulted in the
cloud not being able to update the running machines at all (e.g. even to
shut down the cluster).  This commit makes it so the provider logs a
warning, but returns the machines with as much information as possible.