google: Check whether floating IP is assignable before assigning
Before, if a user specified a floating IP that was not in the user's
project, we would still try to assign it. Because of the way floating
IPs work on Google, this resulted in us removing the existing access
config, and then trying to add the bad IP (which failed). The machine
would then have no public IP at all, rendering it useless.