foreman: Fix frequent tick logic
The foreman has two separate tickers for interacting with the minion:
when the minion is not connected, it tries frequently, and when it is
connected, it tries less often. This code, however, had a bug where it
created a new `connected` variable within the scope of the outer
`connected` variable, rather than assigning the value to the outer
variable. Therefore, the frequent ticker was always used, even if the
minion was already connected.