cloud: Move desiredMachines to cloud.go
This function is now used by join.go and cloud.go, so it makes
sense for it to be in cloud.go rather than with the join-specific
functions in join.go.

Making this change required explicitly setting the admin key in
one of the join tests, because that test assumes the admin key is
empty, and with this move, the admin key will be set (by the test
for desiredMachines) before that join test runs.