google: Use more friendly instance names
This patch removes the kelda prefix, and uses a shorter random number
encoding than UUIDs for google instance names.  This makes them a bit
easier to type in the command line, and more pleasing to read in the
logs. It also reduces our reliance on a random third-party UUID
library.