ci: Block infrastructure setup using test rather than blueprint
Before, we made the infrastructure setup test suite block until the
machines were ready by placing a container that accepts inbound public
connections on every worker. However, this resulted in WARNING logs in the
leader if it tried to schedule the containers before all the workers had
connected to it because it would not think there were enough machines to
run all the containers.

This commit changes the blueprint to simply deploy the machines, and
makes the integration test check whether the machines are all connected.
This will remove the WARNING log message, and also more directly checks
whether the machines are ready.