bindings: Don't choose some machine sizes
When a user doesn't explicitly specify the name of a particular
VM size (e.g, m2.medium), Kelda will choose the cheapest size
that meets the constraints that the user has specified; e.g., if
the user specifies that the machines should have 2 cores, Kelda will
choose the least expensive machine type with 2 cores. This commit
adds an optional flag to the machine descriptions, IgnoredByKelda, that,
if true, specifies that the machine type should not be chosen by Kelda
unless a user explicitly specifies it. The reason behind this is that
some AWS EC2 instance types have unusual semantics: the T2 machines
don't give the user full access to the CPU, and instead limit CPU
use to bursts. It's misleading if we choose this type for a user who
specifies that she'd like, e.g., 1 CPU core, because she can only use
that core in bursts.