supervisor: Refactor to join against actual state
Before, the supervisor code was implemented as a state machine:  it kept
track of whether it had attempted to configure the system for the given
private IP and leader IP, and only updated the state if the IPs were
different. This had several drawbacks:
- It didn't handle failures. If `docker run` failed and a container
didn't actually start, the supervisor would not notice, and a new
container would never be booted.
- The code was hard to follow. It was easy to forget to update the
state. Furthermore, it ignored certain states, such as if the minion
were restarting after a crash, and the system were already partially
configured.
- It was difficult to add new containers to the system. This is
important for the upcoming switch to Kubernetes.

This commit addresses these problems by pulling out the logic for
booting and stopping containers into a single function, and making it so
callers only have to declare the desired containers. When an attribute
such as the leader IP changes, the "desired containers" will change, and
the join function will take care of any necessary changes to the
containers.