kelda: Don't use the global flag object
Before, flags to `kelda` were parsed by modifying the default `FlagSet`
instance in the `flag` package. This meant that if an imported package
configured flags via the global `FlagSet` instance as well, the options
would show up in our `kelda --help` output. This commit avoids this
problem by parsing the flags with a local `FlagSet`.