path: Move MinionTLSDir constant into cli package
An upcoming commit will add another path for the minion. The most
natural place for it is in the `cli/path` package since
`connection/credentials/tls/io` is too specific to TLS. In preparation
for this change, this commit moves the MinionTLSDir constant that was
previously declared in the `tls/io` package into `cli/path`.