ci: Test container restarts by querying container directly
Before, we tested whether containers were restarted when switching
between namespaces by querying Kelda's list of containers, and ensuring
that the container metadata was exactly the same. This commit changes
the test so that the containers serve their start time via a web server
-- if a container restarts, querying the web server will reflect that.

This approach more directly tests whether the container has restarted.
For example, if the container restarted, but a bug in Kelda caused the
container's metadata to not be updated, this test will still properly
fail. The new approach also requires less reasoning about the
implementation of the Kelda scheduler.