engine: Express port placements with hostnames rather than IDs
The commit 6443bedb (scheduler: Write placements in terms of hostnames)
modified the placement API so that placement rules were expressed in
terms of container hostnames rather than blueprint IDs; however, this
patch neglected to update the port placement rule generation to create
those rules in terms of hostnames. This made it possible for containers
listening on the same port to be scheduled on the same machine.