Revert "cloud: Implement preemptible instances on GCE"
This reverts commit 1f5c2c2e19377d916de2d30706efce6e81dec067.

When instances are preempted in Google, they are terminated, but not
actually removed from the GCE project (because the disks are saved,
users can choose to restart them).

Because the instances still show up in the list, Kelda wouldn't start
new VMs to take the place of the preempted ones, so the cluster would
forever be missing some machines.