ovs: Follow redirects when downloading OVS source
For some reason, the OVS release page sometimes responds with a HTTP
redirect rather than directly responding with the source tarball. If we
don't follow the redirects, then the OVS installation script fails
because it doesn't actually download the tarball.

This bug was causing the DigitalOcean provider to fail occasionally
because we don't prebuild the OVS kernel modules for the DigitalOcean
kernel version, requiring the OVS install script to be invoked at
runtime.