debug-logs: Don't exit prematurely if we fail to query cluster
Before, if we were unable to fetch the running machines or containers,
we exited without collecting any logs at all. However, there are many
times when there might be useful logs to be collected. For example, we
might be unable to query the current containers because the master
minion crashed, in which case the machine logs would be very useful.