bindings: Type check arguments to Image constructor
This commit changes the Image constructor to use the `getString` helper
to check that the given arguments are strings.
As a consequence, Image objects will have an empty string rather than `undefined`
as their dockerfile property, when no dockerfile argument is passed. This
causes a number of Container IDs to change in the unit tests.