docs: Update README
This commit updates the README to better reflect the current state of
the project, and to have less text and more code.
Note that this commit switches from using the `bash` language tag to
using the `console` language tag. We do this because `console` will
highlight the commands and output, which `bash` won't.