foreman: Write machine role directly to database
Before, the foreman stored the status information for minions (their
connection state, and their role) in the foreman package, and required
the cloud join to query the roles every time it ran its join.

This commit makes it so that the foreman writes the status information
directly to the database, and makes the cloud join operate on just the
information in the database. This way, the cloud join will trigger on
changes to the role, just like any other database attribute.