kelda: Support running containers in privileged mode
This commit adds the `Privileged` field to the `Container` class to
support running containers in Privileged mode. The effect is the same as
running `docker run --privileged IMAGE` on the command line.

This feature is necessary in order to mount FUSE mounts, since the
container needs access to the host's /dev/fuse device.