digitalocean: Cache List() calls
Digital Ocean has a rather aggressive API request limit that we're
hitting in integration tests.  Part of the problem, is that the API is
global, serving all regions, rather than segregated per region as it
is in other cloud providers.  This means that instead of making a
single List() call that could request all droplets, we were making a
list call per region, artificially inflating our requests per second.
This patch solves the issue by caching the result of a List() call for
1 second, allowing it to be reused by other regions without making API
requests.