network: Allow repeated hostnames in connections
Before, if `allowTraffic` was called with repeated hostnames in the
`From` or `To` set (e.g. `kelda.allowTraffic([blue, blue], [red], 80)`),
the minion would fail to setup the connection. This was because OVSDB
requires address sets have no duplicate items, so it would reject the
call to create the address set for the connection.

This commit makes it so duplicates are filtered before being passed to
OVSDB.