docs: Add class properties in JSDoc
Users might sometimes want to interact with certain properties of the
created JavaScript objects, but we previously didn't document what
properties were available. With this commit we document the properties
that users might want to access or edit. To not confuse users, we
leave out the properties they shouldn't have to interact with.