ci: Increase the permissible delay in build-dockerfile test
Before, the test failed if any containers were started more than 15
seconds apart. While uncommon, it was possible for the containers to
take longer to boot because of unrelated issues, such as network latency.

This commit increases the maximum duration to account for such
issues, while still asserting that the images were built in parallel.